Fix reference to 64-bit integer type.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@7171 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXportLib/include/os/OsTimer.h
@@ -239,9 +239,9 @@ class OsTimer : public UtlContainable
    /**< Class type used for runtime checking */
 
    /// type for absolute time
-   typedef long long int Time;
+   typedef INT64 Time;
    /// type for time interval
-   typedef long long int Interval;
+   typedef INT64 Interval;
 
    OsBSem          mBSem;      //< semaphore to lock access to members
